remove nip07 and webLn hooks
